refactor(use-sorted-classes): rename arbitrary value matcher
