+++ crates/biome_js_analyze/src/lint/nursery/use_sorted_classes/arbitrary_value_match.rs
